Move example into plugins
