+++ src/plugins/extra/mod.rs
@@ -13,19 +13,21 @@
 //! let html = md.parse(r#"Markdown done "The Right Way(TM)""#).render();
 //! assert_eq!(html.trim(), r#"<p>Markdown done “The Right Way™”</p>"#);
 //! ```
-pub mod strikethrough;
-pub mod tables;
 pub mod beautify_links;
+pub mod heading_anchor;
 #[cfg(feature = "linkify")]
 pub mod linkify;
 pub mod smartquotes;
+pub mod strikethrough;
 #[cfg(feature = "syntect")]
 pub mod syntect;
+pub mod tables;
 pub mod typographer;
 
 use crate::MarkdownIt;
 
 pub fn add(md: &mut MarkdownIt) {
+    heading_anchor::add(md);
     strikethrough::add(md);
     beautify_links::add(md);
     #[cfg(feature = "linkify")]
